+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.9] - 2025-ß2-10
+- Callback after every batch in model training now also get gradient norm
+- Removed redundant argument to the YamlWriter
+- Added JsonWriter and JsonReader
+
 ## [0.3.8] - 2025-01-28
 - Added a shorter log format
 - Parent directory of log files will now be created if it does not exist